Added exercises 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -43,6 +43,13 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (blog.user.toString() !== decodedToken.id) {
+    return response
+      .status(401)
+      .json({ error: 'user is not writer of that blog' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
